+++ Network/Telegram/API/Bot/Object/Callback/Notification.hs
@@ -1,0 +1,3 @@
+module Network.Telegram.API.Bot.Object.Callback.Notification (Notification) where
+
+data Notification