+++ Network/Telegram/API/Bot/Object/Notification.hs
@@ -1,3 +1,0 @@
-module Network.Telegram.API.Bot.Object.Notification (Notification) where
-
-data Notification